Refactor navigation and dagger 2, many other changes, debugging functionality

+++ muciplayer-data/src/main/java/com/andreich/data/datasource/remote/RemoteDataSourceImpl.kt
@@ -13,7 +13,7 @@ class RemoteDataSourceImpl(
         return apiService.getChartTracks()
     }
 
-    override suspend fun getTrack(id: Int): TrackDetailDto {
+    override suspend fun getTrack(id: Long): TrackDetailDto {
         return apiService.getTrack(id)
     }
